Fix styles and remove Main.js
