+++ src/index.scss
@@ -7,9 +7,13 @@ body {
   padding: 0;
 }
 
-#root{
+#root {
   width: 100%;
   height: 100%;
+}
+
+#root > .pushable {
+  margin-top: 0;
 }
 
 .padded-item {
@@ -44,15 +48,14 @@ body {
   display: flex !important;
 }
 
-.cern-brand-color{
-  background-color: #0053A1;
+.cern-brand-color {
+  background-color: #0053a1;
   color: #fff;
 }
 
 .cern-brand-color.button {
-  @extend .cern-brand-color
+  @extend .cern-brand-color;
 }
-
 
 ::-webkit-scrollbar {
   width: 5px !important;
@@ -63,15 +66,13 @@ body {
 ::-webkit-scrollbar-track-piece {
 }
 ::-webkit-scrollbar-thumb {
-
 }
 
-
-.button.flat{
+.button.flat {
   background: none;
 }
 
-.button.flat:hover{
+.button.flat:hover {
   background: none;
 }
 
@@ -82,7 +83,6 @@ body {
 //  padding: .78571429em .0em .78571429em;
 //}
 
-
 /* Semantic UI has these classes, however they're only applicable to*/
 /* grids, containers, rows and columns.*/
 /* plus, there isn't any `mobile hidden`, `X hidden` class.*/
@@ -92,60 +92,60 @@ body {
 
 /* Mobile */
 @media only screen and (max-width: 767px) {
-  [class*="mobile hidden"],
-  [class*="tablet only"]:not(.mobile),
-  [class*="computer only"]:not(.mobile),
-  [class*="large screen only"]:not(.mobile),
-  [class*="widescreen only"]:not(.mobile),
-  [class*="or lower hidden"] {
+  [class*='mobile hidden'],
+  [class*='tablet only']:not(.mobile),
+  [class*='computer only']:not(.mobile),
+  [class*='large screen only']:not(.mobile),
+  [class*='widescreen only']:not(.mobile),
+  [class*='or lower hidden'] {
     display: none !important;
   }
 }
 
 /* Tablet / iPad Portrait */
 @media only screen and (min-width: 768px) and (max-width: 991px) {
-  [class*="mobile only"]:not(.tablet),
-  [class*="tablet hidden"],
-  [class*="computer only"]:not(.tablet),
-  [class*="large screen only"]:not(.tablet),
-  [class*="widescreen only"]:not(.tablet),
-  [class*="or lower hidden"]:not(.mobile) {
+  [class*='mobile only']:not(.tablet),
+  [class*='tablet hidden'],
+  [class*='computer only']:not(.tablet),
+  [class*='large screen only']:not(.tablet),
+  [class*='widescreen only']:not(.tablet),
+  [class*='or lower hidden']:not(.mobile) {
     display: none !important;
   }
 }
 
 /* Computer / Desktop / iPad Landscape */
 @media only screen and (min-width: 992px) and (max-width: 1199px) {
-  [class*="mobile only"]:not(.computer),
-  [class*="tablet only"]:not(.computer),
-  [class*="computer hidden"],
-  [class*="large screen only"]:not(.computer),
-  [class*="widescreen only"]:not(.computer),
-  [class*="or lower hidden"]:not(.tablet):not(.mobile) {
+  [class*='mobile only']:not(.computer),
+  [class*='tablet only']:not(.computer),
+  [class*='computer hidden'],
+  [class*='large screen only']:not(.computer),
+  [class*='widescreen only']:not(.computer),
+  [class*='or lower hidden']:not(.tablet):not(.mobile) {
     display: none !important;
   }
 }
 
 /* Large Monitor */
 @media only screen and (min-width: 1200px) and (max-width: 1919px) {
-  [class*="mobile only"]:not([class*="large screen"]),
-  [class*="tablet only"]:not([class*="large screen"]),
-  [class*="computer only"]:not([class*="large screen"]),
-  [class*="large screen hidden"],
-  [class*="widescreen only"]:not([class*="large screen"]),
-  [class*="or lower hidden"]:not(.computer):not(.tablet):not(.mobile) {
+  [class*='mobile only']:not([class*='large screen']),
+  [class*='tablet only']:not([class*='large screen']),
+  [class*='computer only']:not([class*='large screen']),
+  [class*='large screen hidden'],
+  [class*='widescreen only']:not([class*='large screen']),
+  [class*='or lower hidden']:not(.computer):not(.tablet):not(.mobile) {
     display: none !important;
   }
 }
 
 /* Widescreen Monitor */
 @media only screen and (min-width: 1920px) {
-  [class*="mobile only"]:not([class*="widescreen"]),
-  [class*="tablet only"]:not([class*="widescreen"]),
-  [class*="computer only"]:not([class*="widescreen"]),
-  [class*="large screen only"]:not([class*="widescreen"]),
-  [class*="widescreen hidden"],
-  [class*="widescreen or lower hidden"] {
+  [class*='mobile only']:not([class*='widescreen']),
+  [class*='tablet only']:not([class*='widescreen']),
+  [class*='computer only']:not([class*='widescreen']),
+  [class*='large screen only']:not([class*='widescreen']),
+  [class*='widescreen hidden'],
+  [class*='widescreen or lower hidden'] {
     display: none !important;
   }
 }
